socket block size
  * limit sock block size to 8096 for transfer to fldigi
  * add time interval between blocks to allow receiving
    socket to flush it's receive buffers
  * clean up warnings
  * add LOG_INFO statements to transfer function
  * correct computation of time intervals
  * changed FLAMP SCRIPT variable TX_INTERVAL to float
  * update documentation

+++ src/include/status.h
@@ -66,7 +66,7 @@ struct status {
 	bool repeat_forever;
 
 	bool use_txrx_interval;
-	int  tx_interval_minutes;
+	float  tx_interval_minutes;
 	int  rx_interval_seconds;
 
 	bool use_header_modem;
